Add health module and move API implementations to library
The commit includes the creation of a new health module in the API library which houses the implementations for 'hello_world' and 'get_version' endpoints. It also includes the necessary dependencies in 'api-lib' and the reference to it in the main application. The previous implementations of these endpoints in the main application have been removed.